software/src/requirements.txt
pandas==2.2.3
scanpy==1.10.1
anndata==0.10.5
numpy==2.2.6
scipy==1.15.3
scikit-learn==1.6.1
polars-lts-cpu==1.32.1
